pyproject.toml
[tool.poetry]
name = "statistic-bot"
version = "0.0.1"
description = ""
authors = ["Adrian Makridenko <maccree.a@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
lxml = "^4.9.3"
requests = "^2.31.0"
telethon = "^1.29.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
